feat: addLap関数を別ファイルに書き出してimport
